pyproject.toml
[project]
name = "pdf-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "groq>=1.1.2",
    "jsonschema>=4.26.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "ruff>=0.15.8",
    "typer>=0.24.1",
    "uvicorn>=0.42.0",
    "easyocr>=1.7.2",
    "pillow>=11.0.0",
]
